+++ client/src/components/SearchBox.tsx
