+++ client/src/components/news/NewsItem.tsx
