+++ client/src/components/news/NewsItemCompact.css
