+++ client/src/components/news/NewsItemTiny.tsx
