+++ client/src/pages/StyretPage.tsx
